Spinner when map is loading layers
